refactor: move getSizeHandlerPath out of sizlegend
